+++ docs/examples/src/A1_radiation_cylinder.py
@@ -1,20 +1,22 @@
-#!/usr/bin/env python
-
 import numpy as np
 import capytaine as cpt
 
 cpt.set_logging()
 
-# Initialize floating body by generating a geometric mesh
-cylinder = cpt.FloatingBody(
-        mesh=cpt.mesh_horizontal_cylinder(
-            length=10.0, radius=1.0,
-            center=(0, 0, -2),
-            resolution=(10, 20, 30)
-            ))
+# Generating a geometric mesh
+mesh = cpt.mesh_horizontal_cylinder(
+    length=10.0,
+    radius=1.0,
+    center=(0, 0, -2,),
+    resolution=(10, 20, 30,)
+)
 
-# Automatically add the six degrees of freedom of a rigid body
-cylinder.add_all_rigid_body_dofs()
+# Define a rigid body using this mesh
+cylinder = cpt.FloatingBody(
+    mesh=mesh,
+    dofs=cpt.rigid_body_dofs(),
+    name="floating cylinder"
+)
 
 # Define the range of frequencies as a Numpy array
 omega_range = np.linspace(0.1, 3.0, 10)