+++ docs/examples/src/C9_custom_Green_function.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 import numpy as np
 from numpy import dot, pi
 from numpy.linalg import norm