+++ pytest/test_bem_solver.py
@@ -40,6 +40,16 @@ def test_exportable_settings():
     assert solver.exportable_settings['linear_solver'] == 'lu_decomposition'
 
 
+def test_solver_has_initialized_timer():
+    s = cpt.BEMSolver()
+    assert all(t.total == 0.0 for t in s.timer.values())
+
+def test_solver_update_timer(sphere):
+    problem = cpt.DiffractionProblem(body=sphere, omega=1.0)
+    s = cpt.BEMSolver()
+    s.solve(problem)
+    assert not all(t.total == 0.0 for t in s.timer.values())
+
 def test_direct_solver(sphere):
     problem = cpt.DiffractionProblem(body=sphere, omega=1.0)
     solver = cpt.BEMSolver()